move DocumentMessageContent to own class
