+++ samples/apps/copilot-chat-app/webapi/CopilotChat/Controllers/DocumentImportController.cs
@@ -4,8 +4,6 @@ using System;
 using System.Globalization;
 using System.IO;
 using System.Linq;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -44,43 +42,6 @@ public class DocumentImportController : ControllerBase
         /// </summary>
         Pdf,
     };
-
-    /// <summary>
-    /// Value of `Content` for a `ChatMessage` of type `ChatMessageType.Document`.
-    /// </summary>
-    public class DocumentMessageContent
-    {
-        /// <summary>
-        /// Name of the uploaded document.
-        /// </summary>
-        [JsonPropertyName("name")]
-        public string Name { get; set; } = string.Empty;
-
-        /// <summary>
-        /// Size of the uploaded document in bytes.
-        /// </summary>
-        [JsonPropertyName("size")]
-        public string Size { get; set; } = string.Empty;
-
-        /// <summary>
-        /// Serialize the object to a JSON string.
-        /// </summary>
-        /// <returns>A serialized JSON string</returns>
-        public override string ToString()
-        {
-            return JsonSerializer.Serialize(this);
-        }
-
-        /// <summary>
-        /// Deserialize a JSON string to a DocumentMessageContent object.
-        /// </summary>
-        /// <param name="json">A JSON string</param>
-        /// <returns>A DocumentMessageContent object</returns>
-        public static DocumentMessageContent? FromString(string json)
-        {
-            return JsonSerializer.Deserialize<DocumentMessageContent>(json);
-        }
-    }
 
     private readonly ILogger<DocumentImportController> _logger;
     private readonly DocumentMemoryOptions _options;
